requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.70
botocore==1.31.70
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-bootstrap-v5==1.0.11
django-ckeditor==6.6.1
django-cors-headers==4.3.1
django-js-asset==2.1.0
django-phonenumber-field==7.1.0
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
lxml==4.9.3
mysql==0.0.3
mysqlclient
packaging==23.2
phonenumbers==8.13.19
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
s3transfer==0.7.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==6.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4